agent-service/requirements.txt
# CrewAI + AI
crewai[tools]>=0.108.0
crewai-tools>=0.38.0
langchain-anthropic>=0.3.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.2.0

# Database
supabase>=2.13.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0
pydantic>=2.10.0
